Updating the question file
Updated the question file, now we have 15 questions to start with. Currently working on the Master Test Plan and also mapping out what actual functions are required in each system components. I think it might be helpful to use python library Pandas to facilitate easier manipulation of csv and I can start working on that.